refactor add poker template
