+++ codegen.json
@@ -49,6 +49,7 @@
           "ActionMeeting": "../../database/types/MeetingAction#default",
           "ActionMeetingMember": "../../database/types/ActionMeetingMember#default as ActionMeetingMemberDB",
           "AddApprovedOrganizationDomainsSuccess": "./types/AddApprovedOrganizationDomainsSuccess#AddApprovedOrganizationDomainsSuccessSource",
+          "AddPokerTemplateSuccess": "./types/AddPokerTemplateSuccess#AddPokerTemplateSuccessSource",
           "AddTranscriptionBotSuccess": "./types/AddTranscriptionBotSuccess#AddTranscriptionBotSuccessSource",
           "AddedNotification": "./types/AddedNotification#AddedNotificationSource",
           "AgendaItem": "../../database/types/AgendaItem#default as AgendaItemDB",
