+++ packages/server/graphql/public/typeDefs/AddPokerTemplate.graphql
@@ -1,0 +1,31 @@
+extend type Mutation {
+  """
+  Adds a new poker template with a default dimension created.
+  """
+  addPokerTemplate(
+    """
+    The ID of the parent template, if this is a clone operation.
+    """
+    parentTemplateId: ID
+    """
+    The ID of the team for which the template is being created.
+    """
+    teamId: ID!
+  ): AddPokerTemplatePayload!
+}
+
+"""
+Return value for addPokerTemplate, which could be an error
+"""
+union AddPokerTemplatePayload = ErrorPayload | AddPokerTemplateSuccess
+
+type AddPokerTemplateSuccess {
+  """
+  The poker template that was created
+  """
+  pokerTemplate: PokerTemplate!
+  """
+  The user that created the template
+  """
+  user: User!
+}
