+++ packages/server/graphql/rootMutation.ts
@@ -6,7 +6,6 @@ import addComment from './mutations/addComment'
 import addGitHubAuth from './mutations/addGitHubAuth'
 import addIntegrationProvider from './mutations/addIntegrationProvider'
 import addOrg from './mutations/addOrg'
-import addPokerTemplate from './mutations/addPokerTemplate'
 import addPokerTemplateDimension from './mutations/addPokerTemplateDimension'
 import addPokerTemplateScale from './mutations/addPokerTemplateScale'
 import addPokerTemplateScaleValue from './mutations/addPokerTemplateScaleValue'
@@ -129,7 +128,6 @@ export default new GraphQLObjectType<any, GQLContext>({
       addAgendaItem,
       addAtlassianAuth,
       addComment,
-      addPokerTemplate,
       addPokerTemplateDimension,
       addPokerTemplateScale,
       addPokerTemplateScaleValue,
